Fix some circular dependencies
