+++ packages/keystatic/src/app/useItemData.ts
@@ -1,7 +1,7 @@
 import LRUCache from 'lru-cache';
 import { useCallback, useMemo } from 'react';
 import { Config } from '../config';
-import { SlugFieldInfo } from '../form/fields/text/ui';
+import { SlugFieldInfo } from '../form/fields/text/path-slug-context';
 import { ComponentSchema, fields } from '..';
 import { parseProps } from '../form/parse-props';
 import { getAuth } from './auth';
