+++ packages/keystatic/src/form/fields/slug/ui.tsx
@@ -4,7 +4,7 @@ import { Flex, Box } from '@keystar/ui/layout';
 import { TextField } from '@keystar/ui/text-field';
 import { useContext, useState } from 'react';
 import { FormFieldInputProps } from '../../api';
-import { SlugFieldContext, PathContext } from '../text/ui';
+import { SlugFieldContext, PathContext } from '../text/path-slug-context';
 import { validateText } from '../text/validateText';
 
 const emptySet = new Set<string>();
